requirements.txt
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.10.0
EasyProcess==1.1
entrypoint2==1.1
fonttools==4.43.0
idna==2.10
importlib-resources==6.1.0
kiwisolver==1.4.5
matplotlib==3.8.0
MouseInfo==0.1.3
mss==9.0.1
ninja==1.11.1
numpy==1.26.0
opencv-python-headless==4.8.0.74
packaging==23.2
Pillow==9.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
# pyobjc-core==10.0
# pyobjc-framework-Cocoa==10.0
# pyobjc-framework-Quartz==10.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.7
rubicon-objc==0.4.6
scipy==1.11.3
six==1.16.0
supervision==0.14.0
tqdm==4.66.1
urllib3==2.0.5
zipp==3.17.0
